backend/requirements.txt
alembic==1.13.3
Authlib==1.3.2
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-blueprint==1.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
oauthlib==2.1.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.1.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
